fixed registration, but login and ajax-crash

+++ Art_Gallery(CRM)/src/main/java/com/smogorzhevskiy/repository/GalleryRepository.java
@@ -20,7 +20,7 @@ public interface GalleryRepository extends JpaRepository<Gallery,Integer> {
 
     Gallery findOneByName(String name);
 
-    Gallery findOneById(Long id);
+    Gallery findOneById(Integer id);
 //    //language=SQL
 //    @Query("SELECT NAME FROM gallery c WHERE c.name= :name")
 //    Gallery findByName(@Param("name") String name);
